Feat: blog delete by admin

+++ src/app/modules/admin/admin.service.ts
@@ -23,6 +23,13 @@ const blockUserByIdByAdminFromDB = async (
 };
 
 const deleteBlogByIdByAdminFromDB = async (id: string) => {
+  const blog = await Blog.isBlogExistById(id);
+  console.log("blog: ", blog);
+
+  if (!blog) {
+    throw new AppError(StatusCodes.NOT_FOUND, 'Blog not found');
+  }
+
   const result = await Blog.findByIdAndDelete(id);
 
   return result;
